[FIX] 폴더, 문서 이름 구조화 변경
