--- src/components/Card/CardStyle.tsx
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-
-export const BoardWrapper = styled.div`
-  font-size: 30px;
-  background-color: #ffffff;
-  border-radius: 20px;
-  width: 250px;
-  height: 400px;
-  margin: 10px;
-  padding: 10px;
-`;
-
-export const CardWrapper = styled.div`
-  background-color: #dcc573;
-  border-radius: 20px;
-  padding: 5px;
-  text-align: center;
-  font-size: 15px;
-  height: 40px;
-  margin: 5px;
-`;
