+++ src/components/List/ListStyle.tsx
@@ -1,18 +1,21 @@
 import styled from 'styled-components';
 
-export const WorkspaceContentWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-  width: 100%;
-  height: 93vh;
-  min-width: 250px;
-  background-color: ${(props) => props.theme.primaryColor_1};
-  padding: 30px;
+export const ListWrapper = styled.div`
+  font-size: 30px;
+  background-color: #f98080;
+  border-radius: 20px;
+  width: 250px;
+  height: 400px;
+  margin: 10px;
+  padding: 10px;
 `;
 
-export const ColumnDiv = styled.div`
-  display: flex;
-  flex-direction: row;
+export const CardWrapper = styled.div`
+  background-color: #dcc573;
+  border-radius: 20px;
+  padding: 5px;
+  text-align: center;
+  font-size: 15px;
+  height: 40px;
+  margin: 5px;
 `;
