+++ src/pages/Workspace/indexStyle.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
+import boardbackground from '@/assets/images/backgroundimg.jpg';
 
 export const WorkspaceWrapper = styled.div`
   display: flex;
@@ -15,7 +16,7 @@ export const ContentWrapper = styled.div`
   background-color: white;
 `;
 
-export const BoardWrapper = styled.div`
+export const BoardContainer = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
@@ -58,4 +59,36 @@ export const SearchWrapper = styled.div`
   margin: auto 10px;
   height: 1rem;
   border-radius: 10px;
+`;
+
+export const BoardWrapper = styled.div`
+  background-image: url(${boardbackground});
+  background-size: cover;
+  background-repeat: no-repeat;
+  width: 200px;
+  height: 130px;
+  margin: 30px;
+  border-radius: 10px;
+  text-align: center;
+  padding: 10px;
+  font-size: 1rem;
+`;
+
+export const BoardTitle = styled.input`
+  background-color: #ff0000fc;
+  border: 0;
+  text-align: center;
+  width: 180px;
+  height: 30px;
+  border-radius: 5px;
+  font-size: 1rem;
+
+  ::placeholder {
+    color: black;
+  }
+
+  &:focus {
+    background-color: #ffffff;
+    border: 2px solid black;
+  }
 `;
